server: merge nested zone_handler module
